Convert a survey design object to a random-groups jackknife design

Forms a specified number of jackknife replicates based on grouping primary sampling units (PSUs) into random, equal-sized groups. The random-groups jackknife is only valid for designs where the first-stage sampling fraction is ignorable or is identical in every stratum.

Usage

as_random_group_jackknife_design(
  design,
  replicates = 50,
  sampling_fraction = 0,
  group_var_name = ".random_group",
  compress = TRUE,
  mse = getOption("survey.replicates.mse")
)

Arguments

design

A survey design object created using the 'survey' (or 'srvyr') package, with class 'survey.design' or 'svyimputationList'.

replicates

The number of replicates to create. Every stratum must have at least this many primary sampling units (PSUs), or else an error will occur.

sampling_fraction

By default, the random groups jackknife assumes that sampling is with-replacement (i.e., the sampling fraction is ignorable). However, if PSUs were not stratified or if every stratum had the same sampling fraction, you can specify a single value of sampling_fraction, which will be used to adjust the overall replication scale factor through a finite population correction.

group_var_name

(Optional) The name of a new variable created to save identifiers for which random group each PSU was grouped into for the purpose of forming replicates. Specify group_var_name = NULL to avoid creating the variable in the data.

compress

Use a compressed representation of the replicate weights matrix. This reduces the computer memory required to represent the replicate weights and has no impact on estimates.

mse

If TRUE, compute variances from sums of squares around the point estimate from the full-sample weights, If FALSE, compute variances from sums of squares around the mean estimate from the replicate weights.

Value

A replicate design object, with class svyrep.design, which can be used with the usual functions, such as svymean() or svyglm().

Use weights(..., type = 'analysis') to extract the matrix of replicate weights.

Use as_data_frame_with_weights() to convert the design object to a data frame with columns for the full-sample and replicate weights.

Details

The data are sorted by first-stage sampling strata, and then the PSUs in each stratum are randomly arranged. Groups are then formed by sequentially placing observations into each group. The first PSU in the dataset is placed into the first group, the second PSU into the second group, and so on. Once a PSU has been assigned to the last group, the process begins again by assigning the next PSU to the first group, the next PSU to the second group, and so on.

References

- Rust, Keith. 1985. "Variance estimation for complex estimators in sample surveys." Journal of Official Statistics, 1(4), pp.381-397.

- Valliant, Richard, Michael Brick, and Jill Dever. 2008. "Weight adjustments for the grouped jackknife variance estimator." Journal of Official Statistics 24, no. 3: 469.

Examples

library(survey)

# Load example data

 data('api', package = 'survey')

 api_strat_design <- svydesign(
   data = apistrat,
   id = ~ 1,
   strata = ~stype,
   weights = ~pw
 )

# Create a random-groups jackknife design

 jk_design <- as_random_group_jackknife_design(
   api_strat_design,
   replicates = 15
 )
 print(jk_design)
#> Call: as_random_group_jackknife_design(api_strat_design, replicates = 15)
#> with 15 replicates.